requirements.txt
Django==2.1.4
djangorestframework==3.9.0
Pillow==5.3.0
django-filter==2.0.0
psycopg2-binary==2.7.5
django-cors-headers==2.4.0
fabric==2.4.0
celery==4.2.1
django-debug-toolbar==1.11
requests==2.21.0
requests-toolbelt==0.8.0
pytils==0.3
